remove some dependencies of library packages, update package descriptions
