+++ packages/device/package.json
@@ -1,13 +1,21 @@
 {
     "name": "@jaculus/device",
     "version": "0.0.1",
+    "description": "Library for interfacing with Jaculus devices",
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/jaculus-org/Jaculus-tools.git"
+    },
+    "author": {
+        "name": "Petr Kubica (cubicap)"
+    },
+    "license": "GPL-3.0-only",
     "type": "module",
-    "private": false,
+    "main": "./dist/device.js",
+    "types": "./dist/device.d.ts",
     "files": [
         "dist"
     ],
-    "main": "./dist/device.js",
-    "types": "./dist/device.d.ts",
     "scripts": {
         "build": "tsc -b",
         "clean": "rimraf dist tsconfig.tsbuildinfo",
