+++ packages/project/package.json
@@ -1,15 +1,23 @@
 {
     "name": "@jaculus/project",
     "version": "0.0.1",
+    "description": "Library for managing Jaculus projects, including source files, build, and dependencies",
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/jaculus-org/Jaculus-tools.git"
+    },
+    "author": {
+        "name": "Petr Kubica (cubicap)"
+    },
+    "license": "GPL-3.0-only",
     "type": "module",
-    "private": false,
-    "files": [
-        "dist"
-    ],
     "types": "dist/compiler.d.ts",
     "exports": {
         "./compiler": "./dist/compiler.js"
     },
+    "files": [
+        "dist"
+    ],
     "scripts": {
         "build": "tsc -b",
         "clean": "rimraf dist tsconfig.tsbuildinfo",
@@ -20,7 +28,6 @@
         "typescript": "^5.8.3"
     },
     "devDependencies": {
-        "@types/node": "^24.0.7",
         "rimraf": "^6.0.1"
     }
 }
