+++ packages/tools/package.json
@@ -1,7 +1,14 @@
 {
     "name": "jaculus-tools",
     "version": "0.0.26",
-    "description": "Companion CLI for Jaculus",
+    "description": "Companion CLI application for Jaculus",
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/jaculus-org/Jaculus-tools.git"
+    },
+    "author": {
+        "name": "Petr Kubica (cubicap)"
+    },
     "license": "GPL-3.0-only",
     "type": "module",
     "bin": {
@@ -9,6 +16,9 @@
     },
     "main": "dist/cli.js",
     "types": "dist/cli.d.ts",
+    "files": [
+        "dist"
+    ],
     "scripts": {
         "prebuild": "node -p \"'export const version = ' + JSON.stringify(require('./package.json').version) + ';'\" > src/meta/version.ts",
         "build": "tsc -b",
@@ -22,7 +32,6 @@
         "@jaculus/project": "workspace:*",
         "@jaculus/common": "workspace:*",
         "chalk": "^5.4.1",
-        "cli-progress": "^3.12.0",
         "get-uri": "^6.0.4",
         "serialport": "^13.0.0",
         "tar-stream": "^3.1.7",
