+++ packages/tools/src/cli.ts
@@ -3,7 +3,6 @@ import { stdout, stderr } from "process";
 import { Program, Command, Opt } from "./commands/lib/command.js";
 import { registerJaculusCommands } from "./commands/index.js";
 import { logger } from "./logger.js";
-import versionCommand from "./commands/version.js";
 
 const jac = new Program("jac", "Tools for controlling devices running Jaculus", {
     globalOptions: {
@@ -38,11 +37,7 @@ jac.addCommand(
     })
 );
 
-// Reusable command set from package
 registerJaculusCommands(jac);
-
-// App-specific `version`
-jac.addCommand("version", versionCommand);
 
 const args = process.argv.slice(2);
 if (args.length === 0) args.push("help");
